Created a new file for the PoolService.Updated the references in all modules.
