+++ src/app/business/profile/profile.module.ts
@@ -7,7 +7,8 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { ProfileCardComponent } from './profileCard/profile-card.component';
 import { ButtonModule,CardModule,ChartModule,MessageModule,OverlayPanelModule,DialogModule ,ConfirmationService,ConfirmDialogModule, FormModule, DropMenuModule, GrowlModule} from '../../components/common/api';
 import { TooltipModule } from '../../components/tooltip/tooltip';
-import { ProfileService, PoolService } from './profile.service';
+import { ProfileService } from './profile.service';
+import { PoolService } from './pool.service';
 import { HttpService } from '../../shared/api';
 import { SuspensionFrameComponent } from './profileCard/suspension-frame/suspension-frame.component';
